+++ src/test/java/com/alibaba/json/bvt/parser/deser/asm/TestASMEishay.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.asm;
 
 import junit.framework.TestCase;
