+++ src/test/java/com/alibaba/json/bvt/parser/deser/asm/TestASM_boolean.java
@@ -1,31 +1,30 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.asm;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
 
-public class TestASM_int extends TestCase {
+public class TestASM_boolean extends TestCase {
 
     public void test_asm() throws Exception {
         V0 v = new V0();
         String text = JSON.toJSONString(v);
         V0 v1 = JSON.parseObject(text, V0.class);
         
-        Assert.assertEquals(v.getI(), v1.getI());
+        Assert.assertEquals(v.isValue(), v1.isValue());
     }
 
     public static class V0 {
 
-        private int i = 12;
+        private boolean value = true;
 
-        public int getI() {
-            return i;
+        public boolean isValue() {
+            return value;
         }
 
-        public V0 setI(int i) {
-            this.i = i;
-            return this;
+        public void setValue(boolean value) {
+            this.value = value;
         }
 
     }
