+++ src/test/java/com/alibaba/json/bvt/parser/deser/asm/TestASM_double.java
@@ -1,30 +1,30 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.asm;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
 
-public class TestASM_Byte_0 extends TestCase {
+public class TestASM_double extends TestCase {
 
     public void test_asm() throws Exception {
         V0 v = new V0();
         String text = JSON.toJSONString(v);
         V0 v1 = JSON.parseObject(text, V0.class);
-
-        Assert.assertEquals(v.getI(), v1.getI());
+        
+        Assert.assertTrue(v.getValue() == v1.getValue());
     }
 
     public static class V0 {
 
-        private Byte i = 12;
+        private double value = 32.5F;
 
-        public Byte getI() {
-            return i;
+        public double getValue() {
+            return value;
         }
 
-        public void setI(Byte i) {
-            this.i = i;
+        public void setValue(double i) {
+            this.value = i;
         }
 
     }
