+++ src/test/java/com/alibaba/json/bvt/parser/deser/asm/TestASM_int.java
@@ -1,30 +1,31 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.asm;
 
 import org.junit.Assert;
 import junit.framework.TestCase;
 
 import com.alibaba.fastjson.JSON;
 
-public class TestASM_boolean extends TestCase {
+public class TestASM_int extends TestCase {
 
     public void test_asm() throws Exception {
         V0 v = new V0();
         String text = JSON.toJSONString(v);
         V0 v1 = JSON.parseObject(text, V0.class);
         
-        Assert.assertEquals(v.isValue(), v1.isValue());
+        Assert.assertEquals(v.getI(), v1.getI());
     }
 
     public static class V0 {
 
-        private boolean value = true;
+        private int i = 12;
 
-        public boolean isValue() {
-            return value;
+        public int getI() {
+            return i;
         }
 
-        public void setValue(boolean value) {
-            this.value = value;
+        public V0 setI(int i) {
+            this.i = i;
+            return this;
         }
 
     }
