+++ src/test/java/com/alibaba/json/bvt/parser/deser/asm/TestASM_null.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.parser.deser;
+package com.alibaba.json.bvt.parser.deser.asm;
 
 import java.util.List;
